requirements.txt
streamlit
openai
google-generativeai
ffmpeg-python
langchain
faiss-cpu
langchain-community